requirements.txt
Django==4.1.3
colorlog==6.7.0
python-dotenv==0.21.0
pycryptodome==3.14.1